Move watch to cmd/root
