+++ src/cmd/root/watch.ts
